Cambio de int a double para que funcione el monto final

+++ include/Datatypes/DTDetalleCompra.h
@@ -13,7 +13,7 @@ class DTDetalleCompra {
 		DTFecha fechaCompra;
 		map<int, bool> productosEnvio; 
 		set<ParProdCant> productos;
-		DTDetalleCompra(int id, int montoFinal, DTFecha fechaCompra, map<int, bool> productosEnvio, set<ParProdCant> productos);
+		DTDetalleCompra(int id, double montoFinal, DTFecha fechaCompra, map<int, bool> productosEnvio, set<ParProdCant> productos);
 		DTDetalleCompra();
 		~DTDetalleCompra();
 		bool operator<(const DTDetalleCompra &otra) const;
